refactor: web 내의 api, intercepter 내에서 session 불러서 데이터 가져오기
